django_cache/requirements.txt
alabaster==0.7.12
amqp==2.5.1
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
Django==2.2.5
django-celery-beat==1.5.0
django-celerybeat-status==0.0.10
django-mysql==2.2.0
django-nine==0.2.2
django-redis==4.10.0
django-rest-framework-tricks==0.2.10
django-timezone-field==3.0
djangorestframework==3.10.3
docutils==0.15.2
factory-boy==2.12.0
Faker==2.0.1
flower==0.9.3
idna==2.8
imagesize==1.1.0
importlib-metadata==0.20
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.4
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
minepy==1.2.4
mkdocs==1.0.4
more-itertools==7.2.0
mysql-connector-python==8.0.17
mysqlclient==1.4.4
numpy==1.17.1
packaging==19.2
protobuf==3.9.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.4.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
Shapely==1.6.4.post2
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
text-unidecode==1.2
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
